Moved battery
