+++ openconcept/components/__init__.py
@@ -1,2 +1,1 @@
-from .battery import SimpleBattery
 from .manifold import FlowSplit, FlowCombine
