+++ openconcept/energy_storage/__init__.py
@@ -1,0 +1,1 @@
+from .battery import SimpleBattery, SOCBattery
