+++ openconcept/energy_storage/battery.py
