+++ openconcept/examples/propulsion_layouts/simple_series_hybrid.py
@@ -1,7 +1,7 @@
 
 from __future__ import division
 from openconcept.propulsion import SimpleMotor, PowerSplit, SimpleGenerator, SimpleTurboshaft, SimplePropeller
-from openconcept.components.battery import SimpleBattery, SOCBattery
+from openconcept.energy_storage import SimpleBattery, SOCBattery
 from openconcept.atmospherics import ComputeAtmosphericProperties
 from openconcept.utilities.dvlabel import DVLabel
 from openconcept.utilities.math import AddSubtractComp, ElementMultiplyDivideComp
